Break cmd_tlm_server tab gui code into classes
