+++ spec/spec_helper.rb
@@ -12,7 +12,7 @@
 unless ENV['COSMOS_NO_SIMPLECOV']
   require 'simplecov'
   require 'coveralls'
-  Coveralls.wear!
+  Coveralls.wear_merged!
   SimpleCov.formatter = SimpleCov::Formatter::MultiFormatter[
     SimpleCov::Formatter::HTMLFormatter,
     Coveralls::SimpleCov::Formatter
