+++ spec/tools/cmd_tlm_server/router_thread_spec.rb
@@ -22,16 +22,7 @@ module Cosmos
       @packet.buffer = "\x01\x02"
 
       @interface = Interface.new
-      @connected_count = 0
-      allow(@interface).to receive(:connected?) do
-        if @connected_count == 0
-          @connected_count += 1
-          false
-        else
-          @connected_count += 1
-          true
-        end
-      end
+      # Interface#connected? implemented in each test case
       allow(@interface).to receive(:connect)
       allow(@interface).to receive(:disconnect)
       allow(@interface).to receive(:read) do
@@ -40,24 +31,24 @@ module Cosmos
       end
     end
 
-    describe "start" do
-      it "should log the connection" do
-        commanding = double("commanding")
-        expect(commanding).to receive(:send_command_to_interface).at_least(1).times
-        allow(CmdTlmServer).to receive(:commanding).and_return(commanding)
+    describe "handle_packet" do
+      it "handles disconnectd interfaces" do
+        allow(@interface).to receive(:connected?).and_return(false)
         @interface.interfaces = [@interface]
         thread = RouterThread.new(@interface)
-        thread.start
-        sleep 1
-        Thread.list.length.should eql(2)
-        thread.stop
-        sleep 0.5
-        Thread.list.length.should eql(1)
+        capture_io do |stdout|
+          thread.start
+          sleep 0.2
+          Thread.list.length.should eql(2)
+          thread.stop
+          sleep 0.5
+          Thread.list.length.should eql(1)
+          stdout.string.should match "disconnected interface"
+        end
       end
-    end
 
-    describe "handle_packet" do
-      it "should handle errors when sending packets" do
+      it "handles errors when sending packets" do
+        allow(@interface).to receive(:connected?).and_return(true)
         commanding = double("commanding")
         expect(commanding).to receive(:send_command_to_interface).and_raise(RuntimeError.new("Death")).at_least(1).times
         allow(CmdTlmServer).to receive(:commanding).and_return(commanding)
@@ -74,10 +65,43 @@ module Cosmos
         end
       end
 
-      it "should handle identified yet unknown commands" do
+      it "handles identified yet unknown commands" do
+        allow(@interface).to receive(:connected?).and_return(true)
+        commanding = spy("commanding")
+        interface2 = double("interface")
+        allow(interface2).to receive(:connected?).and_return(true)
+        # Setup two interface
+        @interface.interfaces = [@interface, interface2]
+        # Verify that the command gets sent twice (one for each interface)
+        expect(commanding).to receive(:send_command_to_interface).at_least(2).times
+        allow(CmdTlmServer).to receive(:commanding).and_return(commanding)
+        @packet.target_name = 'BOB'
+        @packet.packet_name = 'SMITH'
+        thread = RouterThread.new(@interface)
+        capture_io do |stdout|
+          thread.start
+          sleep 0.2
+          Thread.list.length.should eql(2)
+          thread.stop
+          sleep 0.5
+          Thread.list.length.should eql(1)
+          stdout.string.should match "Received unknown identified command: BOB SMITH"
+        end
+      end
+
+      it "handles identified known commands" do
+        allow(@interface).to receive(:connected?).and_return(true)
         commanding = double("commanding")
+        # Setup two interface
+        interface2 = double("interface")
+        allow(interface2).to receive(:connected?).and_return(true)
+        @interface.interfaces = [@interface, interface2]
+        # Verify that the command gets sent once to the target interface
         expect(commanding).to receive(:send_command_to_interface).at_least(1).times
         allow(CmdTlmServer).to receive(:commanding).and_return(commanding)
+        target = spy("target")
+        allow(target).to receive(:interface).and_return(@interface)
+        allow(System).to receive(:targets).and_return({'BOB' => target})
         @interface.interfaces = [@interface]
         @packet.target_name = 'BOB'
         @packet.packet_name = 'SMITH'
@@ -90,6 +114,35 @@ module Cosmos
           sleep 0.5
           Thread.list.length.should eql(1)
           stdout.string.should match "Received unknown identified command: BOB SMITH"
+        end
+        expect(target).to have_received(:interface).twice #and_return(@interface)
+      end
+
+      it "does not send identified commands with a target and no interface" do
+        allow(@interface).to receive(:connected?).and_return(true)
+        commanding = double("commanding")
+        # Setup two interface
+        interface2 = double("interface")
+        allow(interface2).to receive(:connected?).and_return(true)
+        @interface.interfaces = [@interface, interface2]
+        # Verify that the command gets sent once to the target interface
+        expect(commanding).to_not receive(:send_command_to_interface)
+        allow(CmdTlmServer).to receive(:commanding).and_return(commanding)
+        target = spy("target")
+        allow(target).to receive(:interface).and_return(nil)
+        allow(System).to receive(:targets).and_return({'BOB' => target})
+        @interface.interfaces = [@interface]
+        @packet.target_name = 'BOB'
+        @packet.packet_name = 'SMITH'
+        thread = RouterThread.new(@interface)
+        capture_io do |stdout|
+          thread.start
+          sleep 0.2
+          Thread.list.length.should eql(2)
+          thread.stop
+          sleep 0.5
+          Thread.list.length.should eql(1)
+          stdout.string.should match "target with no interface"
         end
       end
     end
